zero-shot-image-classification/workloads/requirements-oh.txt
sentencepiece
tokenizers
datasets
evaluate
huggingface-hub
transformers==4.50.3
accelerate
peft
diffusers
sentence_transformers
opencv-python
pytesseract
timm
decord
git+https://github.com/facebookresearch/pytorchvideo
requests
librosa
soundfile
pyannote.audio
kenlm
pyctcdecode
Pillow
fire
pytorch_msssim
triton
rouge_score
optimum-habana
